backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
langchain==0.1.4
langchain-openai==0.0.5
langchain-anthropic==0.1.4
python-dotenv==1.0.0
websockets==12.0
